refactor: departure favorites
