+++ src/service/impl/departure-favorites/departure-time.ts
@@ -14,7 +14,7 @@ import {
   GetDepartureRealtimeDocument,
   GetDepartureRealtimeQuery,
   GetDepartureRealtimeQueryVariables
-} from './journey-gql/jp2/departure-time.graphql-gen';
+} from './journey-gql/jp3/departure-time.graphql-gen';
 
 const createVariables = (
   query: DepartureRealtimeQuery
